src/main/java/com/example/order_management_system/service/CustomerService.java
package com.example.order_management_system.service;

import com.example.order_management_system.model.Customer;
import com.example.order_management_system.repository.CustomerRepository;
import org.springframework.stereotype.Service;

@Service
public class CustomerService {

    private final CustomerRepository customerRepository;

    public CustomerService(CustomerRepository customerRepository){
        super();
        this.customerRepository = customerRepository;
    }

    public Customer saveCustomer(Customer customer){
        return customerRepository.save(customer);
    }

}
